refactor(tng-wasm): split fetch module into request/response/attestation
Split the single fetch.rs (~330 lines) into a fetch/ module directory for
clearer separation of concerns:

- mod.rs: public fetch entry, dispatch_request pipeline, forward_request
  (OHTTP layer), and the shared to_js_error helper
- request.rs: parse_request_uri, upstream_endpoint, build_http_request,
  read_body_to_bytes (browser input -> origin-form http::Request)
- response.rs: convert_to_web_response (http::Response -> web_sys::Response)
- attestation.rs: AttestationInfo + attach_attestation_info (RA metadata)

Pure code move; no behavior change. All comments (origin-form/Host rationale,
TODO, AttestationInfo match ladder) preserved verbatim.
