--- tng-wasm/src/fetch.rs
@@ -1,363 +1,0 @@
-use std::sync::Arc;
-
-use anyhow::{bail, Result};
-use tng::{
-    config::{ingress::OHttpArgs, ra::RaArgs},
-    tunnel::{endpoint::TngEndpoint, ingress::protocol::ohttp::security::OHttpSecurityLayer},
-    AttestationResult, RaContext, TokioRuntime,
-};
-
-use anyhow::{anyhow, Context as _};
-use futures::AsyncReadExt;
-use futures::StreamExt;
-use futures::TryStreamExt;
-use gloo::utils::format::JsValueSerdeExt;
-use http_body_util::BodyDataStream;
-use serde::Serialize;
-use tng::config::{
-    ingress::{self},
-    ra::{CocoConverterArgs, CocoVerifierArgs, ConverterArgs, VerifierArgs, VerifyArgs},
-};
-use wasm_bindgen::prelude::*;
-
-#[wasm_bindgen]
-pub async fn fetch(
-    url: String,
-    init: web_sys::RequestInit,
-    config: JsValue,
-) -> Result<web_sys::Response, JsValue> {
-    let common_args: ingress::CommonArgs = serde_wasm_bindgen::from_value(config)
-        .map_err(|e| JsError::new(&format!("Failed to parse config: {e:?}")))?;
-
-    if common_args.web_page_inject {
-        Err(anyhow!("The `web_page_inject` field is not supported"))
-            .map_err(|e| JsError::new(&format!("{e:?}")))?
-    }
-
-    let ohttp = common_args.ohttp.unwrap_or_default();
-    let ra_args = common_args
-        .ra_args
-        .clone()
-        .into_checked()
-        .map_err(|e| JsError::new(&format!("{e:?}")))?;
-
-    let (http_response, attestation_result) = fetch_core_impl(url, init, &ohttp, &ra_args).await?;
-
-    let web_response = convert_to_web_response(http_response).await?;
-    Ok(bind_attestation_result(
-        web_response,
-        attestation_result,
-        &ra_args,
-    )?)
-}
-
-async fn fetch_core_impl(
-    url: String,
-    init: web_sys::RequestInit,
-    ohttp: &OHttpArgs,
-    ra_args: &RaArgs,
-) -> Result<(axum::response::Response, AttestationResult), JsValue> {
-    // Create web_sys::Request
-    let web_request: web_sys::Request = web_sys::Request::new_with_str_and_init(&url, &init)?;
-
-    // Parse the upstream authority from the URL. We can't read it from the
-    // forwarded http::Request's URI because convert_to_rust_request strips
-    // scheme+authority to origin-form (so the BHTTP-encoded request carries
-    // only the path, and the egress gateway emits an origin-form request line
-    // that the backend accepts — an absolute-form line is rejected with 404).
-    let abs_uri: http::Uri = url
-        .parse()
-        .context("Failed to parse request URL")
-        .map_err(|e| JsError::new(&format!("{e:?}")))?;
-    let authority = abs_uri
-        .authority()
-        .context("The authority is empty")
-        .map_err(|e| JsError::new(&format!("{e:?}")))?
-        .clone();
-
-    // TODO: note that in wasm mode, this field should be same as the http request in the body
-    let endpoint = TngEndpoint::new(authority.host(), authority.port_u16().unwrap_or(80));
-
-    let http_request = convert_to_rust_request(web_request).await?;
-
-    send_request_async_impl(&endpoint, ohttp, ra_args, http_request)
-        .await
-        .map_err(|e| JsError::new(&format!("{e:?}")).into())
-}
-
-async fn read_stream_to_vec(
-    body_stream: web_sys::ReadableStream,
-) -> Result<Vec<u8>, anyhow::Error> {
-    let stream = wasm_streams::ReadableStream::from_raw(body_stream)
-        .into_stream()
-        .map(|chunk| {
-            chunk.and_then(|chunk| {
-                let uint8_array = chunk.dyn_into::<js_sys::Uint8Array>().map_err(|e| {
-                    JsValue::from(JsError::new(&format!(
-                        "Failed to convert chunk to Expected Uint8Array: {e:?}"
-                    )))
-                })?;
-                Ok(uint8_array.to_vec())
-            })
-        })
-        .map_err(|e| std::io::Error::other(anyhow::anyhow!("Stream error: {:?}", e)));
-
-    let mut reader = stream.into_async_read(); // This allows using async read methods
-    let mut buffer = Vec::new();
-
-    reader.read_to_end(&mut buffer).await.map_err(|error| {
-        tracing::error!(?error, "Error in read_to_end()");
-        error
-    })?;
-    Ok(buffer)
-}
-
-async fn convert_to_rust_request(
-    web_request: web_sys::Request,
-) -> Result<axum::extract::Request, JsValue> {
-    let gloo_request = gloo::net::http::Request::from(web_request);
-
-    // Get ReadableStream from web_sys::Request and convert to Vec<u8>
-    let body = if let Some(body_stream) = gloo_request.body() {
-        let body_bytes = read_stream_to_vec(body_stream)
-            .await
-            .map_err(|e| JsError::new(&format!("{e:?}")))?;
-        axum::body::Body::from(body_bytes)
-    } else {
-        axum::body::Body::empty()
-    };
-
-    // The browser-built Request carries the full absolute URL but does NOT
-    // expose a `Host` header (fetch sets it internally, outside Headers). The
-    // BHTTP encoder (RFC 9292) encodes scheme+authority when the URI is
-    // absolute, and the egress gateway reconstructs an absolute-form request
-    // line from them — which the PAI-EAS backend rejects (404). The daemon's
-    // http_proxy path forwards an origin-form URI (path+query only) plus a Host
-    // header, which the backend accepts. Mirror that: strip scheme+authority so
-    // BHTTP encodes only the path, and add a Host header derived from the
-    // original authority. (fetch_core_impl parses the authority from the URL
-    // separately for the upstream TngEndpoint, since this URI no longer has it.)
-    let abs_uri: http::Uri = gloo_request
-        .url()
-        .parse()
-        .context("Failed to parse request URL")
-        .map_err(|e| JsError::new(&format!("{e:?}")))?;
-    let authority = abs_uri.authority().map(|a| a.as_str().to_owned());
-    let mut parts = abs_uri.into_parts();
-    parts.scheme = None;
-    parts.authority = None;
-    let origin_uri = http::Uri::from_parts(parts)
-        .context("Failed to convert request URL to origin-form")
-        .map_err(|e| JsError::new(&format!("{e:?}")))?;
-
-    let mut builder = http::Request::builder()
-        .uri(origin_uri)
-        .method(gloo_request.method().as_ref())
-        .version(http::Version::HTTP_11);
-
-    for (key, value) in gloo_request.headers().entries() {
-        builder = builder.header(key, value);
-    }
-    // Ensure a Host header exists (the browser hides it from Headers). Don't
-    // override one the caller set explicitly via init.headers.
-    if let Some(host) = authority {
-        let has_host = builder
-            .headers_ref()
-            .map(|h| h.contains_key(http::header::HOST))
-            .unwrap_or(false);
-        if !has_host {
-            let hv = http::HeaderValue::from_str(&host)
-                .map_err(|e| JsError::new(&format!("invalid host header value: {e:?}")))?;
-            builder = builder.header(http::header::HOST, hv);
-        }
-    }
-
-    let http_request = builder
-        .body(body)
-        .context("Failed to build http::Request")
-        .map_err(|e| JsError::new(&format!("{e:?}")))?;
-
-    Ok(http_request)
-}
-
-async fn convert_to_web_response(
-    http_response: axum::response::Response,
-) -> Result<web_sys::Response, JsValue> {
-    // Convert res.headers() to web_sys::Headers
-    let headers = web_sys::Headers::new()?;
-    for (key, value) in http_response.headers() {
-        let key_str = key.to_string();
-        let value_str = value
-            .to_str()
-            .context("Response header value is not valid UTF-8")
-            .map_err(|e| JsError::new(&format!("{e:?}")))?
-            .to_string();
-        headers.append(&key_str, &value_str)?;
-    }
-
-    // Prepare the ResponseInit
-    let init = web_sys::ResponseInit::new();
-    init.set_headers(&headers);
-    init.set_status(http_response.status().as_u16());
-
-    // Convert body to ReadableStream
-    let body_stream = http_response.into_body();
-    let readable_stream = wasm_streams::ReadableStream::from_stream(
-        BodyDataStream::new(body_stream)
-            .map_ok(|v| {
-                let array = js_sys::Uint8Array::new_with_length(v.len() as u32);
-                array.copy_from(&v);
-                array.into()
-            })
-            .map_err(|e| JsError::new(&format!("{e:?}")).into()),
-    );
-
-    // Create and return the web_sys::Response
-    web_sys::Response::new_with_opt_readable_stream_and_init(
-        Some(&readable_stream.into_raw()),
-        &init,
-    )
-}
-
-/// Non-sensitive subset of verification config exposed to JavaScript.
-///
-/// Only public, non-secret fields are extracted here.  Sensitive values
-/// (`api_key`, `as_headers`, `trusted_certs_paths`, …) are intentionally
-/// omitted.  The explicit match arms ensure a compile error when a new
-/// provider variant is added, forcing a review of which fields to expose.
-#[derive(Serialize)]
-struct AttestationInfo {
-    as_addr: Option<String>,
-    policy_ids: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    as_provider: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    ita_jwks_addr: Option<String>,
-    attestation_result: AttestationResult,
-}
-
-fn bind_attestation_result(
-    web_response: web_sys::Response,
-    attestation_result: AttestationResult,
-    ra_args: &RaArgs,
-) -> Result<web_sys::Response, JsValue> {
-    let mut attest_info = AttestationInfo {
-        as_addr: None,
-        policy_ids: None,
-        as_provider: None,
-        ita_jwks_addr: None,
-        attestation_result,
-    };
-
-    let verify_args = match ra_args {
-        RaArgs::VerifyOnly(v) => Some(v),
-        RaArgs::NoRa => None,
-        #[allow(unreachable_patterns)]
-        _ => None,
-    };
-
-    if let Some(verify_args) = verify_args {
-        match verify_args {
-            VerifyArgs::Passport { verifier } => match verifier {
-                VerifierArgs::Coco(coco) => match coco {
-                    CocoVerifierArgs::Restful {
-                        as_addr,
-                        policy_ids,
-                        ..
-                    }
-                    | CocoVerifierArgs::Grpc {
-                        as_addr,
-                        policy_ids,
-                        ..
-                    } => {
-                        attest_info.as_provider = Some("coco".into());
-                        attest_info.as_addr = as_addr.clone();
-                        attest_info.policy_ids = Some(policy_ids.clone());
-                    }
-                },
-                VerifierArgs::Ita(ita) => {
-                    attest_info.as_provider = Some("ita".into());
-                    attest_info.ita_jwks_addr = Some(ita.ita_jwks_addr.clone());
-                    attest_info.policy_ids = Some(ita.policy_ids.clone());
-                }
-            },
-            VerifyArgs::BackgroundCheck {
-                converter,
-                verifier,
-            } => match converter {
-                ConverterArgs::Coco(coco) => match coco {
-                    CocoConverterArgs::Restful {
-                        as_addr,
-                        policy_ids,
-                        ..
-                    }
-                    | CocoConverterArgs::Grpc {
-                        as_addr,
-                        policy_ids,
-                        ..
-                    } => {
-                        attest_info.as_provider = Some("coco".into());
-                        attest_info.as_addr = Some(as_addr.clone());
-                        attest_info.policy_ids = Some(policy_ids.clone());
-                    }
-                },
-                ConverterArgs::Ita(ita) => {
-                    attest_info.as_provider = Some("ita".into());
-                    attest_info.as_addr = Some(ita.as_addr.clone());
-                    attest_info.policy_ids = Some(ita.policy_ids.clone());
-                    if let VerifierArgs::Ita(v) = verifier {
-                        attest_info.ita_jwks_addr = Some(v.ita_jwks_addr.clone());
-                    }
-                }
-            },
-        }
-    }
-
-    // Create a JavaScript object
-    let attest_info_obj = JsValue::from_serde(&attest_info)
-        .context("Failed to serialize attestation_info object")
-        .map_err(|e| JsError::new(&format!("{e:?}")))?;
-
-    // Set attest_info as a property on the web_response
-    js_sys::Reflect::set(
-        &web_response,
-        &JsValue::from_str("attest_info"),
-        &attest_info_obj,
-    )?;
-
-    Ok(web_response)
-}
-
-async fn send_request_async_impl(
-    endpoint: &TngEndpoint,
-    ohttp: &OHttpArgs,
-    ra_args: &RaArgs,
-    request: axum::extract::Request,
-) -> Result<(axum::response::Response, AttestationResult)> {
-    let request = {
-        let (parts, body) = request.into_parts();
-        tracing::debug!(
-            request=?parts,
-            ?ra_args,
-            "http::Request to be send"
-        );
-        http::Request::from_parts(parts, body)
-    };
-
-    let shutdown = tokio_graceful::Shutdown::no_signal();
-    let runtime = TokioRuntime::wasm_main_thread(shutdown.guard())?;
-
-    let ra_context = Arc::new(RaContext::from_ra_args(&ra_args).await?);
-    let ohttp_security_layer = OHttpSecurityLayer::new(ohttp, ra_context, runtime.clone()).await?;
-
-    let (response, attestation_result) = ohttp_security_layer
-        .forward_http_request(endpoint, request)
-        .await?;
-
-    tracing::info!(?attestation_result, "start forward task");
-    let Some(attestation_result) = attestation_result else {
-        bail!("The attestation result is missing")
-    };
-
-    Ok((response, attestation_result))
-}
